+++ CHANGELOG.md
@@ -1,19 +1,30 @@
 # Note for the future 2.x versions
 
-New developments should use these parameters, to be compliant with default values of the 2.0 version:
+New developments that use __graphql goal/task__ should use the parameters below, to be compliant with default values of the 2.0 version:
 * copyRuntimeSources: false
 * generateDeprecatedRequestResponse: false
 * separateUtilityClasses: true
+
+
+# 1.10
+
+Both mode (client and server):
+* Upgrade of graphql-java from v14.0 to v15.0
+* The main improvement for this is: the plugin now accepts interfaces that implement one or more interfaces
+* Attributes of input types that are enum are now properly managed
+
+Server mode:
+* The generated code would not compile if a GraphQL interface is defined, but not used, in the given GraphQL schema
 
 
 # 1.9
 
 Both mode (client and server):
 * The GraphQL schema can now be split into separate files, including one file containing GraphQL extend keyword on the other file's objects
-* Add of the _merge_ goal/ _mergeTask_ task: it generates a GraphQL schema file, based on the source GraphQL schemas. It can be used to merge several GraphQL schema files into one file, or to reformat the schema files.
+* Add of the _merge_ goal/task: it generates a GraphQL schema file, based on the source GraphQL schemas. It can be used to merge several GraphQL schema files into one file, or to reformat the schema files.
 
 Client mode:
-* Fixes #46 (logged in the GraphQL maven plugin project): Strings not properly escaped for JSON
+* Fixes #46: Strings not properly escaped for JSON
 
 
 # 1.8.1